requirements.txt
crewai==1.6.1
crewai-tools==1.6.1
langchain-groq==0.1.3
# pydantic: allow a modern version compatible with crewai
pydantic>=2.11.9
python-dotenv>=1.1.1